Show warning when comment is already deleted (#27)

+++ src/Controller/App/Review/Comment/DeleteCommentController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace DR\GitCommitNotification\Controller\App\Review\Comment;
 
 use DR\GitCommitNotification\Controller\AbstractController;
+use DR\GitCommitNotification\Controller\App\Review\ProjectsController;
 use DR\GitCommitNotification\Controller\App\Review\ReviewController;
 use DR\GitCommitNotification\Entity\Review\Comment;
 use DR\GitCommitNotification\Repository\Review\CommentRepository;
@@ -24,8 +25,12 @@ class DeleteCommentController extends AbstractController
     #[Route('app/comments/{id<\d+>}', name: self::class, methods: 'DELETE')]
     #[IsGranted(Roles::ROLE_USER)]
     #[Entity('comment')]
-    public function __invoke(Comment $comment): Response
+    public function __invoke(?Comment $comment): Response
     {
+        if ($comment === null) {
+            return $this->refererRedirect(ProjectsController::class, filter: ['action']);
+        }
+
         $this->denyAccessUnlessGranted(CommentVoter::DELETE, $comment);
 
         $this->commentRepository->remove($comment, true);
